CI: move to self-hosted GitHub ARC infrastructure
We are now running the self-hosted testing infrastructure with the help
of GitHub's ARC. This allows us to be more considerate with the
available resources as the VMs are now spawned on demand with the
requested kernel.

For now we have to print KERNEL_VERSION in a previous step to pick up
the correct kernel. This should be fixed in the future by contributing a
labling mechanism to GitHub's ARC.

Signed-off-by: Dennis Maisenbacher <dennis.maisenbacher@wdc.com>